--HG--
branch : dev

+++ Hast.Common/Models/IDeviceManifest.cs
@@ -34,6 +34,6 @@ namespace Hast.Layer
         /// <summary>
         /// Gets the amount of memory (RAM) available to hardware implementations, in bytes.
         /// </summary>
-        uint AvailableMemoryBytes { get; }
+        ulong AvailableMemoryBytes { get; }
     }
 }
